splitting test_timeseries.py further
